+++ src/main/java/janggi/repository/JdbcConnectionManager.java
@@ -1,4 +1,4 @@
-package janggi.persistence;
+package janggi.repository;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
