+++ src/main/java/janggi/service/CsvInitialBoardProvider.java
@@ -1,4 +1,4 @@
-package janggi.application;
+package janggi.service;
 
 import janggi.dto.PositionInfo;
 import janggi.util.FileParser;
