+++ src/main/java/janggi/service/InitialBoardProvider.java
@@ -1,4 +1,4 @@
-package janggi.application;
+package janggi.service;
 
 import janggi.dto.PositionInfo;
 import java.util.List;
