+++ src/test/java/janggi/service/FakeInitialBoardProvider.java
@@ -1,4 +1,4 @@
-package janggi.application;
+package janggi.service;
 
 import janggi.dto.PositionInfo;
 import java.util.List;
